% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatherNetworks2.R
\name{GatherNetworks2}
\alias{GatherNetworks2}
\title{Gather pathway information from RaMP-DB}
\usage{
GatherNetworks2(
  SE,
  IDs = "KEGG",
  database = c("KEGG", "RaMP"),
  species = "hsa",
  minPathwaySize = 5
)
}
\arguments{
\item{SE}{A \code{\link[SummarizedExperiment]{SummarizedExperiment}} with
the features of interest in the 
first \code{\link[SummarizedExperiment]{assay}}.}

\item{IDs}{Character vector indicating the name of the column(s) in \code{SE}
row data containing database IDs}

\item{database}{Character vector indicating the name of the databases you 
wish to query. Options are "KEGG", "RaMP", and "RaMPLocal"}

\item{species}{The three letter KEGG organism ID. If using RaMP, this
argument is ignored.}

\item{minPathwaySize}{Restriction on the minimum number of nodes in networks
returned. Must be greater than or equal to 2.}
}
\value{
a list object containing the original SummarizedExperiment and
igraph network objects
}
\description{
Takes a \code{\link[SummarizedExperiment]{SummarizedExperiment}}
 and constructs a list of KEGG or RaMP pathways along with their associated 
 networks
}
\details{
Queries local RaMP database for known molecular interactions between
included metabolites
}
\examples{
\dontrun{

data(smokers)

# Gather networks from RaMP-DB
networks <- GatherNetworksRaMPLocal(SE = smokers, IDs = "kegg_id",
                                    database = "KEGG",
                                    minPathwaySize = 5)
}

}
