% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatherNetworksRaMPLocal.R
\name{GatherNetworksRaMPLocal}
\alias{GatherNetworksRaMPLocal}
\title{Gather pathway information from RaMP-DB}
\usage{
GatherNetworksRaMPLocal(SE, IDs, minPathwaySize = 5)
}
\arguments{
\item{SE}{A \code{\link[SummarizedExperiment]{SummarizedExperiment}} with
the features of interest in the 
first \code{\link[SummarizedExperiment]{assay}}.}

\item{minPathwaySize}{Filter pathways that are below a minimum size}

\item{ID}{Character vector indicating the name of the column(s) in \code{SE}
row data containing database IDs}
}
\value{
a list object containing the original SummarizedExperiment and
igraph network objects
}
\description{
Takes a \code{\link[SummarizedExperiment]{SummarizedExperiment}}
 and constructs a list of RaMP pathways along with their associated networks
}
\details{
Queries local RaMP database for known molecular interactions between
included metabolites
}
\examples{
\dontrun{

data(smokers)

# Create RaMP environment
pkg.globals <- setConnectionToRaMP(dbname = "ramp2", username = "root", 
                                   conpass = "", host = "localhost")

# Gather networks from RaMP-DB
networks <- GatherNetworksRaMPLocal(SE = smokers, ID = "kegg_id",
                               minPathwaySize = 5)
}


}
