% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatherNetworks.R, R/GatherNetworksRaMP.R
\name{GatherNetworks}
\alias{GatherNetworks}
\alias{GatherNetworksRaMP}
\title{Gather pathway information from KEGG through the KEGGREST API.}
\usage{
GatherNetworks(SE, keggID = "KEGG", species = "hsa", minPathwaySize = 5)

GatherNetworksRaMP(SE, IDs, minPathwaySize = 5)
}
\arguments{
\item{SE}{A \code{\link[SummarizedExperiment]{SummarizedExperiment}} with
the features of interest in the 
first \code{\link[SummarizedExperiment]{assay}}.}

\item{keggID}{column name in pathway data containing KEGG IDs}

\item{species}{The three letter KEGG organism ID}

\item{minPathwaySize}{Filter pathways that are below a minimum size}

\item{ID}{Character vector indicating the name of the column(s) in \code{SE}
row data containing database IDs}
}
\value{
a list object containing the original SummarizedExperiment and
igraph network objects

a list object containing the original SummarizedExperiment and
igraph network objects
}
\description{
Takes a \code{\link[SummarizedExperiment]{SummarizedExperiment}}
 and constructs a list with KEGG pathways

Takes a \code{\link[SummarizedExperiment]{SummarizedExperiment}}
 and constructs a list of RaMP pathways along with their associated networks
}
\details{
Queries KEGG database for known molecular interactions between
included metabolites via the KEGGREST API.

Queries local RaMP database for known molecular interactions between
included metabolites
}
\examples{
data(smokers)
# Query KEGGREST API

networks <- GatherNetworks(SE = smokers, keggID = "kegg_id",
species = "hsa", minPathwaySize = 5)

\dontrun{

data(smokers)

# Create RaMP environment
pkg.globals <- setConnectionToRaMP(dbname = "ramp2", username = "root", 
                                   conpass = "", host = "localhost")

# Gather networks from RaMP-DB
networks <- GatherNetworksRaMP(SE = smokers, ID = "kegg_id",
                               minPathwaySize = 5)
}

}
