% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatherNetworks.R
\name{GatherNetworks2}
\alias{GatherNetworks2}
\alias{GatherNetworks}
\title{Gather pathway information from RaMP-DB}
\usage{
GatherNetworks(
  SE,
  idCols = "KEGG",
  database = c("KEGG", "RaMP"),
  nameCol = idCols[1],
  species = "hsa",
  minPathwaySize = 5
)
}
\arguments{
\item{SE}{A \code{\link[SummarizedExperiment]{SummarizedExperiment}} with
the features of interest in the 
first \code{\link[SummarizedExperiment]{assay}}.}

\item{idCols}{Character vector indicating the name of the column(s) in \code{SE}
row data containing database IDs}

\item{database}{Character vector indicating the name of the databases you 
wish to query. Options are "KEGG", "RaMP", and "RaMPLocal"}

\item{species}{The three letter KEGG organism ID. If using RaMP, this
argument is ignored.}

\item{minPathwaySize}{Restriction on the minimum number of nodes in networks
returned. Must be greater than or equal to 2.}
}
\value{
a list object containing the original SummarizedExperiment and
igraph network objects
}
\description{
Takes a \code{\link[SummarizedExperiment]{SummarizedExperiment}}
 and constructs a list of KEGG or RaMP pathways along with their associated 
 networks
}
\details{
Queries local RaMP database for known molecular interactions between
included metabolites
}
\examples{

data(smokers)

# Gather networks from KEGG
networksKEGG <- GatherNetworks(SE = smokers, idCols = "kegg_id",
                               database = "KEGG",
                               minPathwaySize = 5)

# Gather networks from RaMP
rowData(smokers)$kegg_id <- paste0("kegg:", rowData(smokers)$kegg_id)
networksRaMP <- GatherNetworks(SE = smokers, idCols = "kegg_id",
                               database = "KEGG",
                               minPathwaySize = 5)

}
